requirements.txt
google-generativeai
langchain-google-genai
streamlit
python-dotenv
langchain
gdown
docx2txt